backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4
python-dotenv==1.0.0
websocket-client==1.6.4
lxml==4.9.3
